+++ lib/core/esm.development.js
@@ -1405,7 +1405,7 @@ let reconcileChildren = (current, isRenderRoot) => {
     }
 };
 
-let _reconcileOnlyChildOfDynamicNode = (current, alternate, isRenderRoot) => {   
+let _reconcileOnlyChildOfDynamicNode = (current, alternate, isRenderRoot) => {
     // If the current has an alternate
     // Important note: The alternate of the render root always is null
     if (alternate !== null) {
