+++ tests/src/testcases/update-state-in-effect/unmounted_react.js
