+++ tests/src/testcases/use-effect/effectUnmount_react.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Root} from './effectUnmount.react';
+import {Root} from './effectUnmount_react';
 import {cleanupView, React_renderViewAsync} from '../../../testUtils';
 import {screen} from '@testing-library/dom';
 
@@ -23,82 +23,82 @@ describe('effectUnmount.test.js', () => {
         await sleep();
         expect(onUnmount).toBeCalledTimes(1);
 
-        // console.log('having unmount');
+        // having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(2);
         
-        // console.log('not having unmount');
+        // not having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(2);
 
-        // console.log('not having unmount');
+        // not having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(2);
 
-        // console.log('not having unmount');
+        // not having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(2);
 
-        // console.log('not having unmount');
+        // not having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(2);
 
-        // console.log('having unmount');
+        // having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(3);
 
-        // console.log('having unmount');
+        // having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(4);
 
-        // console.log('having unmount');
+        // having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(5);
 
-        // console.log('not having unmount');
+        // not having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(5);
 
-        // console.log('having unmount');
+        // having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(6);
         
-        // console.log('having unmount');
+        // having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(7);
         
-        // console.log('not having unmount');
+        // not having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(7);
         
-        // console.log('having unmount');
+        // having unmount
         checkbox.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(8);
         
-        // console.log('not having unmount');
+        // not having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(8);
         
-        // console.log('not having unmount');
+        // not having unmount
         button.click();
         await sleep();
         expect(onUnmount).toBeCalledTimes(8);
 
-        // console.log('not having unmount');
+        // not having unmount
     });
 
     afterEach(() => {
